Switch ZFS commands from module to class
To allow properties to be added in the future to ZFS if needed,
such as root dataset, pool, and such, encapsulate all ZFS commands
in a ZFS class. This also acts as a namespace to avoid clashes
with python's builtins, for example 'list' would end up shadowing
python's 'list'.

Where possible all ZFS will be class methods, or static methods.